feat: 修改启动脚本
