+++ apps/common/management/commands/services/hands.py
@@ -1,19 +1,7 @@
 import os
 import sys
-import abc
-import time
-import signal
-import daemon
-import shutil
 import logging
-import datetime
-import threading
-import subprocess
-from daemon import pidfile
-from collections import defaultdict
-from django.db.models import TextChoices
 from django.conf import settings
-from django.core.management.base import BaseCommand, CommandError
 
 from apps.jumpserver.const import CONFIG
 
@@ -28,8 +16,6 @@ except ImportError as e:
     sys.exit(1)
 
 
-WORKERS = 4
-LOG_KEEP_DAYS = 7
 HTTP_HOST = CONFIG.HTTP_BIND_HOST or '127.0.0.1'
 HTTP_PORT = CONFIG.HTTP_LISTEN_PORT or 8080
 WS_PORT = CONFIG.WS_LISTEN_PORT or 8082
@@ -39,3 +25,16 @@ LOG_DIR = os.path.join(BASE_DIR, 'logs')
 APPS_DIR = os.path.join(BASE_DIR, 'apps')
 TMP_DIR = os.path.join(BASE_DIR, 'tmp')
 
+
+"""
+import datetime
+import threading
+import subprocess
+from daemon import pidfile
+from collections import defaultdict
+import abc
+import time
+import signal
+import daemon
+import shutil
+"""
